# openapi.yaml
openapi: 3.0.3
info:
  title: Arepos Server API
  description: REST API для управления моделями, узлами, связями и компонентами
  version: 0.1.1
  contact:
    name: API Support
servers:
  - url: http://arepos-server.arch.svc.cluster.local:8080
    description: Локальный сервер разработки
  - url: https://api.example.com
    description: Продакшн сервер
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/relations:
    get:
      tags:
        - relations
      summary: Список связей
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          schema:
            type: string
        - name: notationId
          in: query
          description: Фильтр по нотации
          schema:
            type: string
            format: uuid
        - name: ownerId
          in: query
          description: Фильтр по владельцу
          schema:
            type: string
            format: uuid
        - name: name
          in: query
          description: Фильтр по имени (поиск по подстроке)
          schema:
            type: string
        - name: tagsAll
          in: query
          description: Фильтр по тегам (в attrs.tags должны присутствовать все перечисленные теги, через запятую)
          schema:
            type: string
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageRelationResponse'
    post:
      tags:
        - relations
      summary: Создать связь
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RelationRequest'
      responses:
        '201':
          description: Связь создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelationResponse'
        '404':
          description: Владелец, нотация или тип связи не найдены
        '400':
          description: Неверный запрос
  /api/v1/relations/{id}:
    get:
      tags:
        - relations
      summary: Получить связь по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelationResponse'
        '404':
          description: Связь не найдена
    put:
      tags:
        - relations
      summary: Обновить связь
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RelationUpdateRequest'
      responses:
        '200':
          description: Связь обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelationResponse'
        '404':
          description: Связь, владелец, нотация или тип связи не найдены
    delete:
      tags:
        - relations
      summary: Удалить связь
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Связь удалена
        '404':
          description: Связь не найдена
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access токен. Получите через /api/v1/auth/login или /api/v1/auth/register
  schemas:
...
    PageInfo:
      type: object
      description: Метаданные пагинации (Spring Data Page)
      properties:
        totalElements:
          type: integer
          format: int64
          description: Общее количество элементов
        totalPages:
          type: integer
          description: Общее количество страниц
        size:
          type: integer
          description: Размер страницы
        number:
          type: integer
          description: Номер текущей страницы (с 0)
        first:
          type: boolean
          description: Первая страница
        last:
          type: boolean
          description: Последняя страница
        numberOfElements:
          type: integer
          description: Количество элементов на текущей странице
        empty:
          type: boolean
          description: Пустая страница
        sort:
          type: object
          nullable: true
          description: Параметры сортировки

    # Пользователи
...
    RelationRequest:
      type: object
      required:
        - name
        - version
        - notationId
        - linkTypeId
      properties:
        name:
          type: string
        version:
          type: string
        notationId:
          type: string
          format: uuid
        ownerId:
          type: string
          format: uuid
          description: Если не указан, берется из JWT токена
        linkTypeId:
          type: string
          format: uuid
        attrs:
          type: string
          nullable: true
    RelationUpdateRequest:
      type: object
      properties:
        name:
          type: string
          nullable: true
        version:
          type: string
          nullable: true
        notationId:
          type: string
          format: uuid
          nullable: true
        ownerId:
          type: string
          format: uuid
          nullable: true
        linkTypeId:
          type: string
          format: uuid
          nullable: true
        attrs:
          type: string
          nullable: true
    RelationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        version:
          type: string
        notationId:
          type: string
          format: uuid
        ownerId:
          type: string
          format: uuid
        linkTypeId:
          type: string
          format: uuid
        attrs:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
          nullable: true
          description: Дата и время создания
        updatedAt:
          type: string
          format: date-time
          nullable: true
          description: Дата и время последнего обновления
    PageRelationResponse:
      allOf:
        - $ref: '#/components/schemas/PageInfo'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/RelationResponse'

    # Компоненты
...